added crud subproduct

See merge request TheoPawelec/8-commerce!40

+++ server/src/Controller/ProductController.php
@@ -18,7 +18,7 @@ use Symfony\Component\Serializer\Exception\NotEncodableValueException;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
-use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 
 class ProductController extends AbstractController
 {
@@ -37,17 +37,18 @@ class ProductController extends AbstractController
      */
     public function productCreate(Request $request, SerializerInterface $serializer, EntityManagerInterface $em, ValidatorInterface $validator)
     {
-        $jsonContent = $request->getContent();
-        $req = json_decode($jsonContent);
-
         try {
-            $product = $serializer->deserialize($jsonContent, Product::class, 'json', [AbstractNormalizer::IGNORED_ATTRIBUTES => ['category', 'subproducts']]);
+            $jsonContent = $request->getContent();
+            $req = json_decode($jsonContent);
+            $product = $serializer->deserialize($jsonContent, Product::class, 'json', [
+                    AbstractNormalizer::IGNORED_ATTRIBUTES => ['category', 'subproducts'],
+                    ObjectNormalizer::DISABLE_TYPE_ENFORCEMENT => true
+                ]);
             $category = $this->getDoctrine()
                 ->getRepository(Category::class)
                 ->find($req->category);
             $product->setCategory($category);
             $product->setCreatedAt(new DateTime());
-
 
             $error = $validator->validate($product);
             if (count($error) > 0) return $this->json($error, 400);
@@ -56,7 +57,6 @@ class ProductController extends AbstractController
             $em->flush();
 
             return $this->json(['product' => $product], 201, [], ['groups' => 'products']);
-
         } catch (NotEncodableValueException $e) {
             return $this->json($e->getMessage(), 400);
         }
@@ -66,10 +66,9 @@ class ProductController extends AbstractController
      * @Route("/api/product/{id}", name="product_details", methods="GET", requirements={"id":"\d+"})
      */
     public function productDetails(Request $request, ProductRepository $productRepository)
-    {  
-        // dd($request->attributes->get('id'));
+    {
         $product = $productRepository->findOneBy(['id' => $request->attributes->get('id')]);
-        if($product){
+        if ($product) {
             return $this->json($product, 200, [], ['groups' => 'products']);
         } else {
             return $this->json(['message' => 'not found'], 404, []);
@@ -77,15 +76,48 @@ class ProductController extends AbstractController
     }
 
     /**
+     * @Route("/api/product/{id}", name="product_update", methods="PUT")
+     */
+    public function productUpdate(Request $request, EntityManagerInterface $em, ValidatorInterface $validator, ProductRepository $productRepository)
+    {
+        try {
+            $jsonContent = $request->getContent();
+            $req = json_decode($jsonContent);
+            $product = $productRepository->findOneBy(['id' => $request->attributes->get('id')]);
+            if(isset($req->category)) $category = $this->getDoctrine()->getRepository(Category::class)->find($req->category);
+                
+            $product->setCategory($category);
+
+            $error = $validator->validate($product);
+            if (count($error) > 0) return $this->json($error, 400);
+
+            $em->persist($product);
+            $em->flush();
+
+            return $this->json(['product' => $product], 200, [], ['groups' => 'products']);
+        } catch (NotEncodableValueException $e) {
+            return $this->json($e->getMessage(), 400);
+        }
+    }
+
+    /**
      * @Route("/api/product/{id}", name="product_remove", methods="DELETE")
      */
-    public function productRemove(Product $product, EntityManagerInterface $em)
+    public function productRemove(Request $request, ProductRepository $productRepository, EntityManagerInterface $em)
     {
-        $em->remove($product);
-        $em->flush();
-        return $this->json([
-            'message' => 'product removed'
-        ], 200);
+        $product = $productRepository->findOneBy(['id' => $request->attributes->get('id')]);
+
+        if ($product) {
+            $em->remove($product);
+            $em->flush();
+
+            return $this->json([
+                'message' => 'product removed',
+                'product' => $product
+            ], 200, [], ['groups' => 'products']);
+        } else {
+            return $this->json(['message' => 'not found'], 404, []);
+        }
     }
 
     /**
